+++ python/cugraph/__init__.py
@@ -13,6 +13,7 @@
 
 from cugraph.community import (
     ecg,
+    ktruss_subgraph,
     louvain,
     spectralBalancedCutClustering,
     spectralModularityMaximizationClustering,
@@ -34,7 +35,7 @@ from cugraph.structure import (
 )
 
 from cugraph.centrality import katz_centrality, betweenness_centrality
-from cugraph.cores import core_number, k_core, ktruss_subgraph
+from cugraph.cores import core_number, k_core
 from cugraph.components import weakly_connected_components, strongly_connected_components
 from cugraph.link_analysis import pagerank