+++ python/cugraph/community/ktruss_subgraph.py
@@ -11,7 +11,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from cugraph.cores import ktruss_subgraph_wrapper
+from cugraph.community import ktruss_subgraph_wrapper
+from cugraph.utilities.unrenumber import unrenumber
+from cugraph.structure.graph import Graph
 
 
 def ktruss_subgraph(G, k, use_weights=True):
@@ -75,9 +77,23 @@ def ktruss_subgraph(G, k, use_weights=True):
     >>> k_subgraph = cugraph.ktruss_subgraph(G, 3)
     """
 
-    KTrussSubgraph = type(G)()
+    KTrussSubgraph = Graph()
+    if type(G) is not Graph:
+        raise Exception("input graph must be undirected")
 
-    ktruss_subgraph_wrapper.ktruss_subgraph(G, k, use_weights,
-                                            KTrussSubgraph)
+    subgraph_df = ktruss_subgraph_wrapper.ktruss_subgraph(G, k, use_weights)
+    if G.renumbered:
+        subgraph_df = unrenumber(G.edgelist.renumber_map, subgraph_df, 'src')
+        subgraph_df = unrenumber(G.edgelist.renumber_map, subgraph_df, 'dst')
+
+    if G.edgelist.weights:
+        KTrussSubgraph.from_cudf_edgelist(subgraph_df,
+                                          source='src',
+                                          destination='dst',
+                                          edge_attr='weight')
+    else:
+        KTrussSubgraph.from_cudf_edgelist(subgraph_df,
+                                          source='src',
+                                          destination='dst')
 
     return KTrussSubgraph